Moved code to the correct files
